Updating dependencies and fixing ios 9 warnings. Reachability and SBJson.

+++ clients/ios/ASI/ASIInputStream.m
@@ -61,7 +61,7 @@ static NSLock *readLock = nil;
 	[readLock unlock];
 	NSInteger rv = [stream read:buffer maxLength:toRead];
 	if (rv > 0)
-		[ASIHTTPRequest incrementBandwidthUsedInLastSecond:rv];
+		[ASIHTTPRequest incrementBandwidthUsedInLastSecond:(NSUInteger)rv];
 	return rv;
 }
